restructure to make way for v2 commands
