+++ internal/command/mpg/detach.go
@@ -7,9 +7,9 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/superfly/flyctl/internal/appconfig"
 	"github.com/superfly/flyctl/internal/command"
+	cmdv1 "github.com/superfly/flyctl/internal/command/mpg/v1"
 	"github.com/superfly/flyctl/internal/flag"
 	"github.com/superfly/flyctl/internal/flyutil"
-	"github.com/superfly/flyctl/internal/uiexutil"
 	"github.com/superfly/flyctl/iostreams"
 )
 
@@ -37,11 +37,6 @@ Note: This does NOT remove any secrets from the app. Use 'fly secrets unset' to 
 }
 
 func runDetach(ctx context.Context) error {
-	// Check token compatibility early
-	if err := validateMPGTokenCompatibility(ctx); err != nil {
-		return err
-	}
-
 	var (
 		clusterId = flag.FirstArg(ctx)
 		appName   = appconfig.NameFromContext(ctx)
@@ -74,17 +69,5 @@ func runDetach(ctx context.Context) error {
 			appName, appOrgSlug, cluster.Id, clusterOrgSlug)
 	}
 
-	uiexClient := uiexutil.ClientFromContext(ctx)
-
-	// Delete the attachment record
-	_, err = uiexClient.DeleteAttachment(ctx, cluster.Id, appName)
-	if err != nil {
-		return fmt.Errorf("failed to detach: %w", err)
-	}
-
-	fmt.Fprintf(io.Out, "\nPostgres cluster %s has been detached from %s\n", cluster.Id, appName)
-	fmt.Fprintf(io.Out, "Note: This only removes the attachment record. Any secrets (like DATABASE_URL) are still set on the app.\n")
-	fmt.Fprintf(io.Out, "Use 'fly secrets unset DATABASE_URL -a %s' to remove the connection string.\n", appName)
-
-	return nil
+	return cmdv1.RunDetach(ctx, cluster.Id, appName)
 }
