+++ internal/command/mpg/proxy.go
@@ -5,14 +5,11 @@ import (
 	"fmt"
 
 	"github.com/spf13/cobra"
-	"github.com/superfly/flyctl/agent"
 	"github.com/superfly/flyctl/internal/command"
+	cmdv1 "github.com/superfly/flyctl/internal/command/mpg/v1"
 	"github.com/superfly/flyctl/internal/flag"
 	"github.com/superfly/flyctl/internal/flag/flagnames"
-	"github.com/superfly/flyctl/internal/flyutil"
-	"github.com/superfly/flyctl/internal/uiex"
 	"github.com/superfly/flyctl/internal/uiexutil"
-	"github.com/superfly/flyctl/proxy"
 )
 
 func newProxy() (cmd *cobra.Command) {
@@ -46,114 +43,32 @@ func newProxy() (cmd *cobra.Command) {
 }
 
 func runProxy(ctx context.Context) (err error) {
-	// Check token compatibility early
-	if err := validateMPGTokenCompatibility(ctx); err != nil {
-		return err
-	}
-
-	localProxyPort := flag.GetString(ctx, flagnames.LocalPort)
-	_, params, _, err := getMpgProxyParams(ctx, localProxyPort, "")
-	if err != nil {
-		return err
-	}
-
-	return proxy.Connect(ctx, params)
-}
-
-func getMpgProxyParams(ctx context.Context, localProxyPort string, username string) (*uiex.ManagedCluster, *proxy.ConnectParams, *uiex.GetManagedClusterCredentialsResponse, error) {
 	clusterID := flag.FirstArg(ctx)
-	var cluster *uiex.ManagedCluster
 	var orgSlug string
-	var err error
 
 	if clusterID != "" {
 		// If cluster ID is provided, fetch directly without prompting for org
 		uiexClient := uiexutil.ClientFromContext(ctx)
 		response, err := uiexClient.GetManagedClusterById(ctx, clusterID)
 		if err != nil {
-			return nil, nil, nil, fmt.Errorf("failed retrieving cluster %s: %w", clusterID, err)
+			return fmt.Errorf("failed retrieving cluster %s: %w", clusterID, err)
 		}
-		cluster = &response.Data
-		orgSlug = cluster.Organization.Slug
+		orgSlug = response.Data.Organization.Slug
 	} else {
 		// Otherwise, prompt for org/cluster selection
-		cluster, orgSlug, err = ClusterFromArgOrSelect(ctx, clusterID, "")
+		cluster, resolvedOrgSlug, err := ClusterFromArgOrSelect(ctx, clusterID, "")
 		if err != nil {
-			return nil, nil, nil, err
+			return err
 		}
+		clusterID = cluster.Id
+		orgSlug = resolvedOrgSlug
 	}
 
-	return getMpgProxyParamsWithCluster(ctx, localProxyPort, username, cluster.Id, orgSlug)
-}
-
-func getMpgProxyParamsWithCluster(ctx context.Context, localProxyPort string, username string, clusterID string, orgSlug string) (*uiex.ManagedCluster, *proxy.ConnectParams, *uiex.GetManagedClusterCredentialsResponse, error) {
-	client := flyutil.ClientFromContext(ctx)
-	uiexClient := uiexutil.ClientFromContext(ctx)
-
-	// Get cluster details
-	response, err := uiexClient.GetManagedClusterById(ctx, clusterID)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("failed retrieving cluster %s: %w", clusterID, err)
-	}
-
-	cluster := &response.Data
-
-	// Get credentials - use user-specific endpoint if username provided, otherwise use default
-	var credentials uiex.GetManagedClusterCredentialsResponse
-	if username != "" {
-		userCreds, err := uiexClient.GetUserCredentials(ctx, cluster.Id, username)
-		if err != nil {
-			return nil, nil, nil, fmt.Errorf("failed retrieving credentials for user %s: %w", username, err)
-		}
-		// Convert user credentials to the standard format
-		credentials = uiex.GetManagedClusterCredentialsResponse{
-			User:     userCreds.Data.User,
-			Password: userCreds.Data.Password,
-			DBName:   response.Credentials.DBName, // Use default DB name from cluster credentials
-		}
-	} else {
-		credentials = response.Credentials
-	}
-
-	// Validate cluster state (only for default credentials, user credentials don't have status)
-	if username == "" {
-		if credentials.Status == "initializing" {
-			return nil, nil, nil, fmt.Errorf("cluster is still initializing, wait a bit more")
-		}
-
-		if credentials.Status == "error" || credentials.Password == "" {
-			return nil, nil, nil, fmt.Errorf("error getting cluster password")
-		}
-	} else if credentials.Password == "" {
-		return nil, nil, nil, fmt.Errorf("error getting user password")
-	}
-
-	if cluster.IpAssignments.Direct == "" {
-		return nil, nil, nil, fmt.Errorf("error getting cluster IP")
-	}
-
-	// Resolve organization slug to handle aliases
+	// Resolve org slug alias for wireguard tunnel
 	resolvedOrgSlug, err := AliasedOrganizationSlug(ctx, orgSlug)
 	if err != nil {
-		return nil, nil, nil, fmt.Errorf("failed to resolve organization slug: %w", err)
+		return fmt.Errorf("failed to resolve organization slug: %w", err)
 	}
 
-	// Establish wireguard tunnel
-	agentclient, err := agent.Establish(ctx, client)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	dialer, err := agentclient.ConnectToTunnel(ctx, resolvedOrgSlug, "", false)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	return cluster, &proxy.ConnectParams{
-		Ports:            []string{localProxyPort, "5432"},
-		OrganizationSlug: resolvedOrgSlug,
-		Dialer:           dialer,
-		BindAddr:         flag.GetBindAddr(ctx),
-		RemoteHost:       cluster.IpAssignments.Direct,
-	}, &credentials, nil
+	return cmdv1.RunProxy(ctx, clusterID, resolvedOrgSlug)
 }
