+++ internal/command/mpg/restore.go
@@ -2,14 +2,11 @@ package mpg
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/spf13/cobra"
 	"github.com/superfly/flyctl/internal/command"
+	cmdv1 "github.com/superfly/flyctl/internal/command/mpg/v1"
 	"github.com/superfly/flyctl/internal/flag"
-	"github.com/superfly/flyctl/internal/uiex"
-	"github.com/superfly/flyctl/internal/uiexutil"
-	"github.com/superfly/flyctl/iostreams"
 )
 
 func newRestore() *cobra.Command {
@@ -36,14 +33,6 @@ func newRestore() *cobra.Command {
 }
 
 func runRestore(ctx context.Context) error {
-	// Check token compatibility early
-	if err := validateMPGTokenCompatibility(ctx); err != nil {
-		return err
-	}
-
-	out := iostreams.FromContext(ctx).Out
-	uiexClient := uiexutil.ClientFromContext(ctx)
-
 	clusterID := flag.FirstArg(ctx)
 	if clusterID == "" {
 		cluster, _, err := ClusterFromArgOrSelect(ctx, clusterID, "")
@@ -54,25 +43,5 @@ func runRestore(ctx context.Context) error {
 		clusterID = cluster.Id
 	}
 
-	backupID := flag.GetString(ctx, "backup-id")
-	if backupID == "" {
-		return fmt.Errorf("--backup-id flag is required")
-	}
-
-	fmt.Fprintf(out, "Restoring cluster %s from backup %s...\n", clusterID, backupID)
-
-	input := uiex.RestoreManagedClusterBackupInput{
-		BackupId: backupID,
-	}
-
-	response, err := uiexClient.RestoreManagedClusterBackup(ctx, clusterID, input)
-	if err != nil {
-		return fmt.Errorf("failed to restore backup: %w", err)
-	}
-
-	fmt.Fprintf(out, "Restore initiated successfully!\n")
-	fmt.Fprintf(out, "  Cluster ID: %s\n", response.Data.Id)
-	fmt.Fprintf(out, "  Cluster Name: %s\n", response.Data.Name)
-
-	return nil
+	return cmdv1.RunRestore(ctx, clusterID)
 }
