+++ internal/command/mpg/v1/run_databases.go
@@ -1,0 +1,80 @@
+package cmdv1
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/superfly/flyctl/internal/config"
+	"github.com/superfly/flyctl/internal/flag"
+	"github.com/superfly/flyctl/internal/prompt"
+	"github.com/superfly/flyctl/internal/render"
+	"github.com/superfly/flyctl/internal/uiex"
+	"github.com/superfly/flyctl/internal/uiexutil"
+	"github.com/superfly/flyctl/iostreams"
+)
+
+func RunDatabasesList(ctx context.Context, clusterID string) error {
+	cfg := config.FromContext(ctx)
+	out := iostreams.FromContext(ctx).Out
+	uiexClient := uiexutil.ClientFromContext(ctx)
+
+	databases, err := uiexClient.ListDatabases(ctx, clusterID)
+	if err != nil {
+		return fmt.Errorf("failed to list databases for cluster %s: %w", clusterID, err)
+	}
+
+	if len(databases.Data) == 0 {
+		fmt.Fprintf(out, "No databases found for cluster %s\n", clusterID)
+
+		return nil
+	}
+
+	if cfg.JSONOutput {
+		return render.JSON(out, databases.Data)
+	}
+
+	rows := make([][]string, 0, len(databases.Data))
+	for _, db := range databases.Data {
+		rows = append(rows, []string{
+			db.Name,
+		})
+	}
+
+	return render.Table(out, "", rows, "Name")
+}
+
+func RunDatabasesCreate(ctx context.Context, clusterID string) error {
+	out := iostreams.FromContext(ctx).Out
+	uiexClient := uiexutil.ClientFromContext(ctx)
+
+	dbName := flag.GetString(ctx, "name")
+	if dbName == "" {
+		io := iostreams.FromContext(ctx)
+		if !io.IsInteractive() {
+			return prompt.NonInteractiveError("database name must be specified with --name flag when not running interactively")
+		}
+		err := prompt.String(ctx, &dbName, "Enter database name:", "", true)
+		if err != nil {
+			return err
+		}
+		if dbName == "" {
+			return fmt.Errorf("database name cannot be empty")
+		}
+	}
+
+	fmt.Fprintf(out, "Creating database %s in cluster %s...\n", dbName, clusterID)
+
+	input := uiex.CreateDatabaseInput{
+		Name: dbName,
+	}
+
+	response, err := uiexClient.CreateDatabase(ctx, clusterID, input)
+	if err != nil {
+		return fmt.Errorf("failed to create database: %w", err)
+	}
+
+	fmt.Fprintf(out, "Database created successfully!\n")
+	fmt.Fprintf(out, "  Name: %s\n", response.Data.Name)
+
+	return nil
+}
